+++ tests/teal/router/questionableFP_clear_v8.teal
@@ -1,3 +1,3 @@
 #pragma version 8
-int 0
+int 1
 return
