+++ tests/teal/router/questionable_clear_v6.teal
@@ -1,3 +1,3 @@
 #pragma version 6
-int 0
+int 1
 return
