+++ tests/teal/router/sim_approval_yacc_6.teal
@@ -1,67 +1,42 @@
 #pragma version 6
-txn NumAppArgs
-int 0
-==
-bnz main_l22
 txna ApplicationArgs 0
 method "add(uint64,uint64)uint64"
 ==
-bnz main_l21
+bnz main_l18
 txna ApplicationArgs 0
 method "sub(uint64,uint64)uint64"
 ==
-bnz main_l20
+bnz main_l17
 txna ApplicationArgs 0
 method "mul(uint64,uint64)uint64"
 ==
-bnz main_l19
+bnz main_l16
 txna ApplicationArgs 0
 method "div(uint64,uint64)uint64"
 ==
-bnz main_l18
+bnz main_l15
 txna ApplicationArgs 0
 method "mod(uint64,uint64)uint64"
 ==
-bnz main_l17
+bnz main_l14
 txna ApplicationArgs 0
 method "all_laid_to_args(uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64)uint64"
 ==
-bnz main_l16
+bnz main_l13
 txna ApplicationArgs 0
 method "empty_return_subroutine()void"
 ==
-bnz main_l15
+bnz main_l12
 txna ApplicationArgs 0
 method "log_1()uint64"
 ==
-bnz main_l14
+bnz main_l11
 txna ApplicationArgs 0
 method "log_creation()string"
 ==
-bnz main_l13
-txna ApplicationArgs 0
-method "approve_if_odd(uint32)void"
-==
-bnz main_l12
+bnz main_l10
 err
-main_l12:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-assert
-txna ApplicationArgs 1
-int 0
-extract_uint32
-store 69
-load 69
-callsub approveifodd_9
-int 1
-return
-main_l13:
+main_l10:
 txn OnCompletion
 int NoOp
 ==
@@ -71,14 +46,14 @@ int 0
 &&
 assert
 callsub logcreation_8
-store 67
+store 49
 byte 0x151f7c75
-load 67
+load 49
 concat
 log
 int 1
 return
-main_l14:
+main_l11:
 txn OnCompletion
 int NoOp
 ==
@@ -96,15 +71,15 @@ int 0
 ||
 assert
 callsub log1_7
-store 65
+store 48
 byte 0x151f7c75
-load 65
+load 48
 itob
 concat
 log
 int 1
 return
-main_l15:
+main_l12:
 txn OnCompletion
 int NoOp
 ==
@@ -120,7 +95,7 @@ assert
 callsub emptyreturnsubroutine_6
 int 1
 return
-main_l16:
+main_l13:
 txn OnCompletion
 int NoOp
 ==
@@ -131,46 +106,46 @@ int 0
 assert
 txna ApplicationArgs 1
 btoi
-store 30
+store 22
 txna ApplicationArgs 2
 btoi
-store 31
+store 23
 txna ApplicationArgs 3
 btoi
-store 32
+store 25
 txna ApplicationArgs 4
 btoi
-store 33
+store 27
 txna ApplicationArgs 5
 btoi
-store 34
+store 28
 txna ApplicationArgs 6
 btoi
-store 35
+store 29
 txna ApplicationArgs 7
 btoi
-store 36
+store 31
 txna ApplicationArgs 8
 btoi
-store 37
+store 33
 txna ApplicationArgs 9
 btoi
-store 38
+store 35
 txna ApplicationArgs 10
 btoi
-store 39
+store 36
 txna ApplicationArgs 11
 btoi
-store 40
+store 37
 txna ApplicationArgs 12
 btoi
-store 41
+store 38
 txna ApplicationArgs 13
 btoi
-store 42
+store 39
 txna ApplicationArgs 14
 btoi
-store 43
+store 41
 txna ApplicationArgs 15
 store 46
 load 46
@@ -181,20 +156,20 @@ load 46
 int 8
 extract_uint64
 store 45
-load 30
+load 22
+load 23
+load 25
+load 27
+load 28
+load 29
 load 31
-load 32
 load 33
-load 34
 load 35
 load 36
 load 37
 load 38
 load 39
-load 40
 load 41
-load 42
-load 43
 load 44
 load 45
 callsub alllaidtoargs_5
@@ -206,33 +181,7 @@ concat
 log
 int 1
 return
-main_l17:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-assert
-txna ApplicationArgs 1
-btoi
-store 24
-txna ApplicationArgs 2
-btoi
-store 25
-load 24
-load 25
-callsub mod_4
-store 26
-byte 0x151f7c75
-load 26
-itob
-concat
-log
-int 1
-return
-main_l18:
+main_l14:
 txn OnCompletion
 int NoOp
 ==
@@ -249,7 +198,7 @@ btoi
 store 19
 load 18
 load 19
-callsub div_3
+callsub mod_4
 store 20
 byte 0x151f7c75
 load 20
@@ -258,7 +207,7 @@ concat
 log
 int 1
 return
-main_l19:
+main_l15:
 txn OnCompletion
 int NoOp
 ==
@@ -269,22 +218,22 @@ int 0
 assert
 txna ApplicationArgs 1
 btoi
-store 12
+store 13
 txna ApplicationArgs 2
 btoi
-store 13
-load 12
-load 13
-callsub mul_2
 store 14
-byte 0x151f7c75
+load 13
 load 14
+callsub div_3
+store 16
+byte 0x151f7c75
+load 16
 itob
 concat
 log
 int 1
 return
-main_l20:
+main_l16:
 txn OnCompletion
 int NoOp
 ==
@@ -295,12 +244,38 @@ int 0
 assert
 txna ApplicationArgs 1
 btoi
-store 6
+store 9
 txna ApplicationArgs 2
 btoi
-store 7
+store 10
+load 9
+load 10
+callsub mul_2
+store 11
+byte 0x151f7c75
+load 11
+itob
+concat
+log
+int 1
+return
+main_l17:
+txn OnCompletion
+int NoOp
+==
+txn ApplicationID
+int 0
+!=
+&&
+assert
+txna ApplicationArgs 1
+btoi
+store 3
+txna ApplicationArgs 2
+btoi
+store 6
+load 3
 load 6
-load 7
 callsub sub_1
 store 8
 byte 0x151f7c75
@@ -310,7 +285,7 @@ concat
 log
 int 1
 return
-main_l21:
+main_l18:
 txn OnCompletion
 int NoOp
 ==
@@ -336,79 +311,66 @@ concat
 log
 int 1
 return
-main_l22:
-txn OnCompletion
-int OptIn
-==
-bnz main_l24
-err
-main_l24:
-txn ApplicationID
-int 0
-!=
-assert
-byte "optin call"
-log
-int 1
-return
 
 // add
 add_0:
-store 4
-store 3
-load 3
-load 4
-+
 store 5
+store 4
+load 4
 load 5
++
+store 7
+load 7
 retsub
 
 // sub
 sub_1:
-store 10
-store 9
-load 9
-load 10
--
-store 11
-load 11
-retsub
-
-// mul
-mul_2:
-store 16
 store 15
+store 12
+load 12
 load 15
-load 16
-*
+-
 store 17
 load 17
 retsub
 
-// div
-div_3:
-store 22
+// mul
+mul_2:
+store 24
 store 21
 load 21
-load 22
+load 24
+*
+store 26
+load 26
+retsub
+
+// div
+div_3:
+store 32
+store 30
+load 30
+load 32
 /
-store 23
-load 23
+store 34
+load 34
 retsub
 
 // mod
 mod_4:
-store 28
-store 27
-load 27
-load 28
+store 42
+store 40
+load 40
+load 42
 %
-store 29
-load 29
+store 43
+load 43
 retsub
 
 // all_laid_to_args
 alllaidtoargs_5:
+store 65
+store 64
 store 63
 store 62
 store 61
@@ -423,13 +385,7 @@ store 53
 store 52
 store 51
 store 50
-store 49
-store 48
-load 48
-load 49
-+
 load 50
-+
 load 51
 +
 load 52
@@ -456,8 +412,12 @@ load 62
 +
 load 63
 +
-store 64
 load 64
++
+load 65
++
+store 66
+load 66
 retsub
 
 // empty_return_subroutine
@@ -469,8 +429,8 @@ retsub
 // log_1
 log1_7:
 int 1
-store 66
-load 66
+store 67
+load 67
 retsub
 
 // log_creation
@@ -479,16 +439,3 @@ byte 0x00106c6f6767696e67206372656174696f6e
 store 68
 load 68
 retsub
-
-// approve_if_odd
-approveifodd_9:
-store 70
-load 70
-int 2
-%
-bnz approveifodd_9_l2
-int 0
-return
-approveifodd_9_l2:
-int 1
-return
