+++ tests/teal/router/sim_clear_questionable_6.teal
@@ -1,3 +1,3 @@
 #pragma version 6
-int 0
+int 1
 return
