+++ tests/teal/router/sim_clear_yacc_6.teal
@@ -1,0 +1,3 @@
+#pragma version 6
+int 1
+return
