+++ tests/teal/router/sim_clear_yacc_8.teal
@@ -1,0 +1,3 @@
+#pragma version 8
+int 1
+return
