+++ tests/teal/router/yaccFP_clear_v8.teal
@@ -1,0 +1,3 @@
+#pragma version 8
+int 1
+return
