+++ tests/teal/router/yacc_clear_v6.teal
@@ -1,0 +1,3 @@
+#pragma version 6
+int 1
+return
